+++ src/store/swapStore.ts
@@ -32,6 +32,7 @@ type SwapState = {
   isApproving: boolean;
   strategy: string;
   tokenPrices: TokenPrices;
+  fiatTokenPrices: TokenPrices;
   error: SwapErrors;
   isNetworkFeesLoading: boolean;
   isFetchingQuote: FetchingQuote;
@@ -44,6 +45,7 @@ type SwapState = {
   };
   maxTimeSaved: number;
   maxCostSaved: number;
+  setFiatTokenPrices: (fiatTokenPrices: TokenPrices) => void;
   setTokenPrices: (tokenPrices: TokenPrices) => void;
   setIsSwapping: (isSwapping: boolean) => void;
   setIsApproving: (isApproving: boolean) => void;
@@ -97,6 +99,10 @@ export const swapStore = create<SwapState>((set) => ({
     input: "0",
     output: "0",
   },
+  fiatTokenPrices: {
+    input: "0",
+    output: "0",
+  },
   error: {
     inputError: Errors.none,
     outputError: Errors.none,
@@ -121,6 +127,9 @@ export const swapStore = create<SwapState>((set) => ({
       ...state,
       [ioType === IOType.input ? "inputAsset" : "outputAsset"]: asset,
     }));
+  },
+  setFiatTokenPrices: (fiatTokenPrices) => {
+    set({fiatTokenPrices});
   },
   setAmount: (ioType, amount) => {
     set((state) => ({
@@ -242,6 +251,10 @@ export const swapStore = create<SwapState>((set) => ({
         input: "0",
         output: "0",
       },
+      fiatTokenPrices: {
+        input: "0",
+        output: "0",
+      },
       error: {
         inputError: Errors.none,
         outputError: Errors.none,
@@ -277,6 +290,10 @@ export const swapStore = create<SwapState>((set) => ({
         input: "0",
         output: "0",
       },
+      fiatTokenPrices: {
+        input: "0",
+        output: "0",
+      },
       error: {
         inputError: Errors.none,
         outputError: Errors.none,
@@ -303,6 +320,10 @@ export const swapStore = create<SwapState>((set) => ({
       outputAmount: "",
       rate: 0,
       tokenPrices: {
+        input: "0",
+        output: "0",
+      },
+      fiatTokenPrices: {
         input: "0",
         output: "0",
       },