+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated balance retrieval function to use fallback RPC clients as default instead of default wallet RPCs.
 - Rate conversion and display estimated token price for inputAsset.
+- Fixed fiat token prices.
 
 
 ## [1.1.4]